# src/main/resources/functional_requirements/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/ui/gl-batch": {
      "get": {
        "tags": [
          "gl-batch-controller"
        ],
        "operationId": "getAllGLBatches",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntityWithMetadataGLBatch"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "gl-batch-controller"
        ],
        "operationId": "createGLBatch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GLBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityWithMetadataGLBatch"
                }
              }
            }
          }
        }
      }
    },
    "/ui/accrual": {
      "get": {
        "tags": [
          "accrual-controller"
        ],
        "operationId": "getAllAccruals",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntityWithMetadataAccrual"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "accrual-controller"
        ],
        "operationId": "createAccrual",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Accrual"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityWithMetadataAccrual"
                }
              }
            }
          }
        }
      }
    },
...
    "/ui/gl-batch/{id}": {
      "get": {
        "tags": [
          "gl-batch-controller"
        ],
        "operationId": "getGLBatchById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityWithMetadataGLBatch"
                }
              }
            }
          }
        }
      }
    },
    "/ui/accrual/{id}": {
      "get": {
        "tags": [
          "accrual-controller"
        ],
        "operationId": "getAccrualById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityWithMetadataAccrual"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "EntityMetadata": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "modelKey": {
            "$ref": "#/components/schemas/ModelSpec"
          },
          "state": {
            "type": "string"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "transitionForLatestSave": {
            "type": "string"
          }
        }
      },
...
      "ModelSpec": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "GLApproval": {
        "type": "object",
        "properties": {
          "preparedBy": {
            "type": "string"
          },
          "preparedAt": {
            "type": "string",
            "format": "date-time"
          },
          "approvedBy": {
            "type": "string"
          },
          "approvedAt": {
            "type": "string",
            "format": "date-time"
          },
          "approverRole": {
            "type": "string"
          },
          "approvalComments": {
            "type": "string"
          }
        }
      },
      "GLBatch": {
        "type": "object",
        "properties": {
          "glBatchId": {
            "type": "string"
          },
          "periodStart": {
            "type": "string",
            "format": "date"
          },
          "periodEnd": {
            "type": "string",
            "format": "date"
          },
          "periodCode": {
            "type": "string"
          },
          "totalEntries": {
            "type": "integer",
            "format": "int32"
          },
          "totalDebits": {
            "type": "number"
          },
          "totalCredits": {
            "type": "number"
          },
          "balanceDifference": {
            "type": "number"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GLEntry"
            }
          },
          "exportFileName": {
            "type": "string"
          },
          "exportFormat": {
            "type": "string"
          },
          "exportPath": {
            "type": "string"
          },
          "exportedAt": {
            "type": "string",
            "format": "date-time"
          },
          "glSystemBatchId": {
            "type": "string"
          },
          "postedToGLAt": {
            "type": "string",
            "format": "date-time"
          },
          "glAcknowledgmentId": {
            "type": "string"
          },
          "approval": {
            "$ref": "#/components/schemas/GLApproval"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string"
          },
          "updatedBy": {
            "type": "string"
          }
        }
      },
      "GLEntry": {
        "type": "object",
        "properties": {
          "entryId": {
            "type": "string"
          },
          "accountCode": {
            "type": "string"
          },
          "accountName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "debitAmount": {
            "type": "number"
          },
          "creditAmount": {
            "type": "number"
          },
          "dimension1": {
            "type": "string"
          },
          "dimension2": {
            "type": "string"
          },
          "dimension3": {
            "type": "string"
          },
          "reference": {
            "type": "string"
          }
        }
      },
      "EntityWithMetadataGLBatch": {
        "type": "object",
        "properties": {
          "entity": {
            "$ref": "#/components/schemas/GLBatch"
          },
          "meta": {
            "$ref": "#/components/schemas/EntityMetadata"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "type": "string"
          },
          "modelKey": {
            "$ref": "#/components/schemas/ModelSpec"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "transitionForLatestSave": {
            "type": "string"
          }
        }
      },
      "Accrual": {
        "type": "object",
        "properties": {
          "accrualId": {
            "type": "string"
          },
          "loanId": {
            "type": "string"
          },
          "accrualDate": {
            "type": "string",
            "format": "date"
          },
          "principalBase": {
            "type": "number"
          },
          "apr": {
            "type": "number"
          },
          "dayCountBasis": {
            "type": "string"
          },
          "dayCountNumerator": {
            "type": "integer",
            "format": "int32"
          },
          "dayCountDenominator": {
            "type": "integer",
            "format": "int32"
          },
          "interestAmount": {
            "type": "number"
          },
          "dailyRate": {
            "type": "number"
          },
          "isRecomputation": {
            "type": "boolean"
          },
          "originalAccrualId": {
            "type": "string"
          },
          "recomputationReason": {
            "type": "string"
          },
          "scheduledAt": {
            "type": "string",
            "format": "date-time"
          },
          "calculatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "recordedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string"
          },
          "updatedBy": {
            "type": "string"
          }
        }
      },
      "EntityWithMetadataAccrual": {
        "type": "object",
        "properties": {
          "entity": {
            "$ref": "#/components/schemas/Accrual"
          },
          "meta": {
            "$ref": "#/components/schemas/EntityMetadata"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "type": "string"
          },
          "modelKey": {
            "$ref": "#/components/schemas/ModelSpec"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "transitionForLatestSave": {
            "type": "string"
          }
        }
      }
    }
  }
}